Move history buttons.
