+++ packages/story-editor/src/components/header/buttons/index.js
@@ -16,5 +16,7 @@
 
 export { default as PreviewButton } from './preview';
 export { default as PublishButton } from './publish';
+export { default as RedoButton } from './redo';
 export { default as SwitchToDraftButton } from './switchToDraft';
 export { default as UpdateButton } from './update';
+export { default as UndoButton } from './undo';
